Create Data_Abstraction.py
In this code, Data Abstraction is demonstrated through the use of abstract base classes and derived classes. The Shape class defines the common interface, while the Rectangle and Circle classes provide specific implementations of the area calculation. Objects are instantiated and their area methods are called to obtain the respective areas